Abort if calling TLS macros with gcc<4.4 and -fPIC
There are issues with using TLS the way that I do if you don't have features
present in gcc>=4.4 when building for position independant code (PIC).

Because of this, I now check at configure time if you are compiling with
gcc>=4.4 and only emit a static binary if so.

Before I was issuing a compile time warning everytime a particualar header was
included, as well as forcing a compile time error if executables were built
that called macros where there would be a problem.  Now I'm allowing even
programs that will have problems to be built, but issuing a runtime error and
recommending they rebuild with the -static flag.