Mixed up logic for EWOULDAGAIN in futex_wait()
Apparently futex_wait() returns -1 with an EWOULDAGAIN if the call just falls
through without blocking.  This will only happen if our *uaddr == val
condition has become false before calling it.  In my opinion this should be
a successful case of returning once the condition is broke (whether we blocked
or not), but it seems we disagree once again...