Prime the tsc_freq() for a given vcore.
This calibrates the get_tsc_freq() call on a given core so that, so that
subsequent calls to this function will be fast (they use the cached
value).  The idea is that since a vcore is pinned to an actual physical
core, this value on't change from the first time it is run.