Add new vcore api call vcore_request_specific
This call allows you to rerequest a specific vcore from the system.
This is useful when you want to yield the core back so as not to waste cpu
cycles, but you want to make sure you get the same one back when you reequest
it again in the future (for purposes of per-vcore structures).