Set state back to RUNNING when restart aftre sig
We need to treat being interrupted by a signal as if our current uthread
simply got interrupted.  Part of this leaving current_uthread set.  On akaros
we are actually interrupted and dropped directly into vcore context.  On linux
we actually do a voluntary yield in the signal handler to drop us into vcore
context.  This voluntary yield changes the uthread state to NOT_RUNNING, so we
have to force it back to running before restarting the vcore. This is actually
a benign error because this state is only used for debugging, but I caught it
once when asserts were turned on, so it makes sense to fix it.