Need to define defaults for __real* functions
Since socket and accept cannot easily be wrapped in libc, we have to force an
application to use the -Wl,wrap functionality from ld in order to wrap them.
In order to link to aprlib without using -Wl,wrap, we need default
implementations for the __real* versionos of the functions being wrapped.
This became an issue, for example, when trying to configure upthread because
it does a check that it can create an executable that links against parlib
without any extra options.