No need for interupt_safe here anymore
Previously, calls to uthread_set_tls_var needed to swap out the tls
temporarily in order to set the variable. In order to do this safely, we
needed to disable notifs for any uthread that might have triggered the
uthread_init() call. Since we no longer require a temporary tls change (since
we decalre all our tls variables as TLS_INITIAL_EXEC, and rely on a relative
offset of the current tls base to set them), we no longer need to disable
notifs in uthread_init().