Move defs of __real* into another translation unit
As it says in the manual for the example fo wrapping malloc:
You may wish to provide a __real_malloc function as well, so that links without
the --wrap option will succeed. If you do this, you should not put the
definition of __real_malloc in the same file as __wrap_malloc; if you do, the
assembler may resolve the call before the linker has a chance to wrap it to
malloc.