Fix race in setting arg->tcb
Previously, there was a race with the the value of this variable being
set, and the futex waiting on it in __create_backing_pthread() since it
might never go to sleep.  We only want to actually set arc->tcb right
before waking the futex, otherwise we run the risk of
__creat_backing_pthread() exiting early and the memory for arg->tcb
disappearing out from under us.