Change the way vcores are reserved for allocation
I can't explain what was wrong with the previous implementation, but
under heavy contention, the accounting of __num_vcores would get out of
whack.  Since we used this count as our reservation system for vcores,
we would end up in deadlock because we would succeed in reserving a
vcore even though all vcores were already allocated, and we would get
stuck forever in the loop trying to wake it up.

This previous method was nice because reservations were agnostic to
individual cores, and a reservation could be made / canceled with a
simple cas.

This new method is less efficient, but there is less chance for error.