Fix uth_disable/enable semantics around hijack_uthread
When hijacking a uthread, you need to make sure that you disable and
then reenable notifs for BOTH the original uthread, and the hijacked
uthread.  If you don't, you end up in a weird state for one of the other
of them when trying to disable notifications at the end of this call.

If you don't disable notifs for the new uthread at the beginning, then trying to
reenable them at the end will result in an assertion failure.
Likewise, if you don't reenable notifs for the original uthread after
it's been hijacked, then when we switch back to it via a subsequent call
to hijack_uthread, it's disable depth will be off by one.