Add a list to avoid mmapping sigstacks every time
I wanted to use a wfl, but internally it uses malloc, so that choice is
out.  Instead, every vcore maintains a list of unused stacks that can
set as the sigstack when we need to pin the current one to a uthread.
A new stack is mmapped when the list is empty, otherwise an old one from
the list is used.  Stacks mmapped on one vcore may end up being freed
onto the list of a different vcore, but this shouldn't be a problem.