Initialize a hijacked uthread properly.
There is no guarantee that the uthread passed into this function has
been initialized. Do so now before completing the hijack.  I discovered
this problem because there were times that the uthread had an invalid
sigstack and when it tried to free it in uthread_yield() we were
segfaulting. By initializing it properly here, we avoid these problems.